% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avg_phgrofit.R
\name{avg_phgrofit}
\alias{avg_phgrofit}
\title{avg_phgrofit
Copyright (c) 2019. Kaleido Biosciences. All Rights Reserved}
\usage{
avg_phgrofit(phgrofit_output, group_by = c("Sample.ID"))
}
\arguments{
\item{phgrofit_output}{The output of phgropro.}

\item{group_by}{A character vector of the names of the columns that you would like to group by}
}
\value{
a data frame with mean values for each of the modeling parameters across the groups that were specified.
}
\description{
This function is intended to make averaging the model parameters from phgrofit across groups very easy.
New Sample.IDs are assigned by concatanating the conditions that the user chooses to group by in order to make it easy to match the output avg_phgropro to avg_phgrofit.
}
