% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avg_phgropro.R
\name{avg_phgropro}
\alias{avg_phgropro}
\title{avg_phgropro
Copyright (c) 2019. Kaleido Biosciences. All Rights Reserved}
\usage{
avg_phgropro(phgropro_output, group_by = "Sample.ID")
}
\arguments{
\item{phgropro_output}{The output of phgropro.}

\item{group_by}{Character vector of the names of the columns
that you would like to group by.}
}
\value{
a data frame with mean_OD600, sd_OD600, mean_pH, sd_pH across the groups specified
}
\description{
This function is intended to make averaging OD600 and pH values from phgropro
across groups very easy.
This functiuon also returns the standard deviation for OD600 and pH.
New Sample.IDs are assigned by concatenating the conditions
that the user chooses to group by in order to make it easy to match
the output avg_phgropro to avg_phgrofit.
}
\examples{
#This would return a data frame with the average and sd pH and OD600 values when grouped by Community and Compound.
\dontrun{
p1 = avg_phgropro(phgropro_output,c("Community","Compound"))
}
}
