% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R, R/colored_dendro.R
\name{colored_dendro}
\alias{colored_dendro}
\title{colored_dendro}
\usage{
colored_dendro(annotated_data, cols_with_color_label,
  num_of_clusters = 0, legend_name = "Community")

colored_dendro(annotated_data, cols_with_color_label,
  num_of_clusters = 0, legend_name = "Community")
}
\arguments{
\item{annotated_data}{this is modeling data from phgrobiome that has been combined with relevant metadata through the meta_combine function.}

\item{cols_with_color_label}{this is a vector of the names of the columns that you would like to color on the}

\item{num_of_clusters}{this is the number of clusters that you would like to highlight by boxing. The default is 0}

\item{legend_name}{this is the name of the condition that you would like the legend to be colored by. The defalut legend is Community.
Currently only one legend is supported}

\item{annotated_data}{this is modeling data from phgrobiome that has been combined with relevant metadata through the meta_combine function.}

\item{cols_with_color_label}{this is a vector of the names of the columns that you would like to color on the}

\item{num_of_clusters}{this is the number of clusters that you would like to highlight by boxing. The default is 0}

\item{legend_name}{this is the name of the condition that you would like the legend to be colored by. The defalut legend is Community.
Currently only one legend is supported}
}
\value{
plot of colored dendrogram

plot of colored dendrogram
}
\description{
This function creates a dendrogram using the output from the meta_combine function using the manhatten distance as the distance method.
The ploted dendogram will have a colored bar representing different categorical features of the data as specified by the cols_with_color_label parameter
A specified number of clusters can be highlighted via boxes by the num_of clusters parameter.
This function can plot a colored legend for a categorical feature as specified by the user.

This function creates a dendrogram using the output from the meta_combine function using the manhatten distance as the distance method.
The ploted dendogram will have a colored bar representing different categorical features of the data as specified by the cols_with_color_label parameter
A specified number of clusters can be highlighted via boxes by the num_of clusters parameter.
This function can plot a colored legend for a categorical feature as specified by the user.
}
\details{
For a more detailed and interactive representation of the data, it is suggested to use the heatmapper function.

For a more detailed and interactive representation of the data, it is suggested to use the heatmapper function.
}
\examples{

annotated_data = meta_combine(phgrobiome_output,metadata,compound_info)

colored_dendro(annotated_data,cols_with_color_label = c("Community","Compound","Composition","Manual_Curation"),
num_of_clusters = 3,legend_name = "Composition")

This will result in a colored dendrogram with colored bars representing distinct Communities, Compounds, Composition of compounds,
and a manual curation. The legend will show the corresponding colors for all compositions.

annotated_data = meta_combine(phgrobiome_output,metadata,compound_info)

colored_dendro(annotated_data,cols_with_color_label = c("Community","Compound","Composition","Manual_Curation"),
num_of_clusters = 3,legend_name = "Composition")

This will result in a colored dendrogram with colored bars representing distinct Communities, Compounds, Composition of compounds,
and a manual curation. The legend will show the corresponding colors for all compositions.
}
