% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combined_dendro_plot.R
\name{combined_dendro_plot}
\alias{combined_dendro_plot}
\title{'combined dendro plot
Copyright (c) 2019. Kaleido Biosciences. All Rights Reserved}
\usage{
combined_dendro_plot(dend, cluster_data, colored_bar, k, colored_bar_label)
}
\arguments{
\item{dend}{dendrogram}

\item{cluster_data}{clustering data}

\item{colored_bar}{colored bar}

\item{k}{what is the k?}

\item{colored_bar_label}{what is the label of the colored bar}
}
\value{
a dendrogram plot
}
\description{
'combined dendro plot
Copyright (c) 2019. Kaleido Biosciences. All Rights Reserved
}
\examples{
\dontrun{
p1 = combined_dendro_plot(dend,cluster_data,colored_bar,k,colored_bar_label)
}
}
