% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dendrospect.R
\name{dendrospect}
\alias{dendrospect}
\title{dendrospect
Copyright (c) 2019. Kaleido Biosciences. All Rights Reserved}
\usage{
dendrospect(phgrofit_data, phgropro_data, colored_bar_label = NULL, k = 8)
}
\arguments{
\item{phgrofit_data}{This is the data origniating from phgrofit or one of it's modifying functions such as scale_phgrofit or avg_phgrofit.}

\item{phgropro_data}{This is the data origniating from phgropro or one of it's modifying functions such as avg_phgropro. It must contain the same Sample.IDs as phgrofit data.}

\item{colored_bar_label}{This is the character name of a column that the user would like to include as colored bar below the dendogram.}

\item{k}{The number of clusters.}
}
\value{
A ggplot2 object containin a colored dendogram whith a colored bar label and the corresponding kinetic OD600 and pH data.
}
\description{
Allows the user to easily see what the average kinetic profile corresponding with a given cluster is.
Note that data should be scaled before using this function.
}
\examples{

# phgropro processing
\dontrun{phgropro_output = phgropro(biotek_export = filepath.txt,metadata = metadata.csv,Plate_Type = 96)}

# phgrofit processing
\dontrun{phgrofit_output = phgrofit(phgropro_output)}

## processing phgrofit data by averaging and scaling
\dontrun{ phgrofit_data = avg_phgrofti(phgrofit_output,c("Community","Compound")) \%>\%
scale_phgrofit()}
### plotting heatmap with colored labels for community and mouse over information about the compounds.
\dontrun{plot = PCA(phgrofit_data,"Community")
plot}
}
