% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dendrospect_kinetic.R
\name{dendrospect_kinetic}
\alias{dendrospect_kinetic}
\title{dendrospect_kinetic
Copyright (c) 2019. Kaleido Biosciences. All Rights Reserved}
\usage{
dendrospect_kinetic(phgropro_data, phgrofit_data, k = 8)
}
\arguments{
\item{phgropro_data}{This is the data origniating from phgropro or one of it's modifying functions such as avg_phgropro. It must contain the same Sample.IDs as phgrofit data.}

\item{phgrofit_data}{This is the data origniating from phgrofit or one of it's modifying functions such as scale_phgrofit or avg_phgrofit.}

\item{k:}{The number of clusters.}
}
\value{
a dataframe containing the phgropro data with a column named dendrogram_cluster that represents the cluster the observations correspond to.
}
\description{
Allows the user to easily acess the OD600 and pH kinetic data corresponding with a given cluster.
This is intended to allow for easier examination of the underlying data that is produced with the dendrospect function.
Note that the phgrofit_data should first be scaled before using this function. This can be acomplished with scale_phgrofit() function.
}
\examples{
#Returns a dataframe mapping which of the 8 clusters a Sample.ID belongs to.
df = dendrospect_model(phgropro_data,k = 8)
}
