% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flag.R
\name{flag}
\alias{flag}
\title{flag
Copyright (c) 2019. Kaleido Biosciences. All Rights Reserved}
\usage{
flag(phgrofit_output)
}
\arguments{
\item{phgrofit_output}{a data frane with modeling information from phgrofit}
}
\value{
a data frame with a five columns corresponding to symptoms that may be indicative of poor model fit or weird growth curves
\itemize{
 \item{"negligable growth"}{ TRUE if max OD600 is less than 1.5x the starting od600}
 \item{"negative_lag_length"}{ TRUE if the od600_lag_length was calculated to be negative}
 \item{"starting_od600_higher_than_max"}{ TRUE if the starting od600 is higher than max od600}
 \item{"missing_od600_data"}{ TRUE if 25 percent of the od600 data for a well is missing}
 \item{"missing_pH_data"}{ TRUE if 25 percent of the pH data for a well is missing}
}
}
\description{
This function was added in order to allow the user to easily flag wells that may have had problematic model fitting.
}
\examples{
\dontrun{dataframe_exploring_potential_problems = flag(phgrofit_output)}
# if there are no problems an empty data frame will be returned.
}
