% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_check.R
\name{graph_check}
\alias{graph_check}
\title{graph_check
Copyright (c) 2019. Kaleido Biosciences. All Rights Reserved}
\usage{
graph_check(data)
}
\arguments{
\item{data:}{This is the OD600 + pH data from a single well. Contains Columns for Sample.ID, Time, OD600, and pH.}
}
\value{
a ggplot2 object
}
\description{
This function is used in the exported function model_fit_check to generate a graphical representation of the model fit
}
\examples{

for( i in Sample.ID){
loop_data = data \%>\%
dplyr::filter(Sample.ID == i )
# graph check is set up to plot each well independantly
print(graph_check(loop_data))
}
}
