% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grofit.R
\name{grofit}
\alias{grofit}
\title{grofit
Allows the user to get modeling data for just OD600}
\usage{
grofit(gropro_output)
}
\arguments{
\item{gropro_output}{: This is the output of gropro. This is a tidy data frame that has a column for Sample.ID and OD600.
OD600 must be the only measurement that occurs at each time. Other metadata can be included as necessary.}
}
\value{
:Modeling
}
\description{
grofit
Allows the user to get modeling data for just OD600
}
\examples{

#This will return modeling information for each Sample.ID
grofit_output = grofit(gropro_output)
}
