% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gropro.R
\name{gropro}
\alias{gropro}
\title{gropro: Formatting raw OD600 data from the biotek reader into a tidy data frame.}
\usage{
gropro(biotek_export, metadata, Plate_Type = 96)
}
\arguments{
\item{biotek_export}{export from biotek plate reader}

\item{metadata}{metadata}

\item{Plate_Type}{type of plate 384 or 96}
}
\value{
tidy data frame
}
\description{
gropro: Formatting raw OD600 data from the biotek reader into a tidy data frame.
}
\examples{
#Need to format OD600 data exported from biotek plate reader in an experiment with a 96 well plate

\dontrun{gropro_output = gropro("path_to_biotek_OD600_export", "path_to_metadata.csv",Plate_Type = 96)}

}
