% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmapper.R
\name{heatmapper}
\alias{heatmapper}
\title{heatmapper: Creating an interactive heatmap with colored categorical labels.}
\usage{
heatmapper(annotated_data, labels = "Community")
}
\arguments{
\item{annotated_data}{This is the annotated data that results from merging the modeling data with relevant metadata using the meta_combine function.}

\item{labels}{This is a vector specifiying what colored labels you would like to be displayed beside the heatmap}
}
\value{
a plotly heatmap via heatmaply.
}
\description{
heatmapper: Creating an interactive heatmap with colored categorical labels.
}
