% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_combine.R
\name{meta_combine}
\alias{meta_combine}
\title{meta_combine: Combining modeling data,experimental metadata, and compound information.}
\usage{
meta_combine(phgrobiome_output, metadata, compound_info)
}
\arguments{
\item{phgrobiome_output}{This is a data frame that is loaded into R that contains the Sample.IDs and all of the values resulting from the phgrobiome modeling.
This is the output of the phgrobiome function}

\item{metadata}{This is a data frame that has been loaded into R containing metadata from the experiment.
This datatable contains at least columns labeled Sample.ID, Community, Compound, Compound_Concentration, and Media.}

\item{compound_info}{This is a data table that has been loaded into R that contains information about compounds.
This data table contains at least a column labeled Compound, Composition,AveDP, and Manual_Curation.
Manual_Curation represents a user created column that specifies some sort of note about the compounds.
I found this to be necessary because the compound information in LIMS seemed to be lacking/ missing important categorical information.}
}
\value{
a dataframe containing modeling data, experimental metadata, and compound information.
}
\description{
The purpose of this function is to provide a convienent method of combining the model output of phgrobiome with metadata and compound information.
In doing this, we create an easily manipulatable data frame containing all relacent parameters.
This data frame can then be used with the phgrofit::colored_dendro function or for any other type of analysis that the user wants.
}
