% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_fit_check.R
\name{model_fit_check}
\alias{model_fit_check}
\title{model_fit_check
Copyright (c) 2019. Kaleido Biosciences. All Rights Reserved}
\usage{
model_fit_check(phgropro_output, grouping_vars = "Sample.ID")
}
\arguments{
\item{phgropro_output}{This is the output from phgropro. It contains tidy pH and OD600 data.}

\item{grouping_vars}{This contains the variables you would like to see the fit for a randomly sampled replicate of.}
}
\value{
prints a randomly sampled plot from each condition to the console as specified by grouping_vars.
}
\description{
This function prints graphs visually displaying the model fits from a randomly sampled set of variables of the users choosing.
A replicate from each unique condition specified is randomly sampled and the fit and extracted parameters that are easy to visualize are shown.
}
\examples{
phgropro_output = phgrofit::phgropro_output("Filepath of biotek export.txt","filepath of metadata.csv,Plate_Type = 384)
model_fit_check(phgropro_output,grouping_vars = c("Community","Compound))
#This would print graphs from a randomly sampled replicate of each combination of variables specified by grouping_vars
}
