% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA.R
\name{PCA}
\alias{PCA}
\title{PCA: Easily making a PCA plot with phgrofit data}
\usage{
PCA(phgrofit_output, group = "Sample.ID")
}
\arguments{
\item{phgrofit_output}{this is output from the phgrofit}

\item{group}{this is the factor that you would like to color the ordination by. A 95 percent confidence interval for this group will be assigned.}
}
\value{
a ggplot2 object containing the 95 percent confidence interval for any group specfied in gorup.
}
\description{
PCA: Easily making a PCA plot with phgrofit data
}
\examples{
### phgropro processing
phgropro_output = phgropro(biotek_export = filepath.txt,metadata = metadata.csv,Plate_Type = 96)

### phgrofit processing
phgrofit_output = phgrofit(phgropro_output)

### printing PCA plot with colored confidence intervals for community
community_PCA = PCA(phgrofit_output,"Community")
print(community_PCA)
}
