% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phgrofit.R
\name{phgrofit}
\alias{phgrofit}
\title{Extract physiological parameters from kinetic pH and OD600 data}
\usage{
phgrofit(data, graphs = 1)
}
\arguments{
\item{data}{This is the input data that you would like to model. It will most often be the output off phgropro.}

\item{graphs}{This is the number specifying how many graphs you would like to print to console. This is useful for visually inspecting the modeling.}
}
\value{
The output pf
}
\description{
phgrofit takes pH and OD600 data that has been formated by phgropro and applies a spline interpolation to extract relevant physiological data.
}
