% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phgropro.R
\name{phgropro}
\alias{phgropro}
\title{Tidying data from biotek plate reader}
\usage{
phgropro(biotek_export, Plate_Type = 96)
}
\arguments{
\item{biotek_export}{.txt file that results from exporting data from the Biotek Gen5 software.}

\item{Plate_Type}{96 or 384 specifying which plate type was ran on the plate reader.}
}
\description{
phgropro- pH growh processing- takes the export from a standardized biotek plate reader and converts it into a tidy format that is convenient for data analysis. This tidy format serves as the input to pgrofit
}
\examples{
phgropro(data = filepath,Plate_Type = 96)
}
