% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_pH.R
\name{predict_pH}
\alias{predict_pH}
\title{predict_pH : function to predict pH from standard curve.}
\usage{
predict_pH(standard_data, prediction_data)
}
\arguments{
\item{standard_data}{data frame containing standard curve data. Has to have a column named IR and column named pH}

\item{prediction_data}{data frame with the prediction data, must contain the IR. Must have a column named IR.}
}
\value{
a data frame with predicted pH values
}
\description{
predict_pH : function to predict pH from standard curve.
}
\examples{

standard_data = data.frame(IR = c(1,1.2,1.3,1.5,1.8,2.5,3.5,4.2,4.7,4.8,6),
pH = seq(from =4, to = 9,by = 0.5))
prediction_data = data.frame(time = 1:10, IR = c(5,5,4,4,3,2,2,1.1,4,5))

predicted_pH_data = predict_pH(standard_data,prediction_data)

}
