% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplify.R
\name{simplify}
\alias{simplify}
\title{simplify:Simplifying the model fit by taking the averages across distinct groups.}
\usage{
simplify(annotated_data, groups = c("Community", "Composition"))
}
\arguments{
\item{annotated_data}{This is the annotated data (output of meta_combine) that the user wants to average replicates across.}

\item{groups}{This is a vector input containing the names of the columns that the data will be grouped by}
}
\value{
A data frame containing the averaged summerized model fits across the groups provided by the user.
}
\description{
simplify:Simplifying the model fit by taking the averages across distinct groups.
}
\examples{
This code assigns the averaged model fit across Communities, Manual Curations, and Compositions to the object "summarized data".
sumarized_data = simplify(annotated_data,groups = c("Community","Manual Curation","Composition"))


}
