% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combined_dendro_plot.R
\name{combined_dendro_plot}
\alias{combined_dendro_plot}
\title{Copyright (c) 2019. Kaleido Biosciences. All Rights Reserved}
\usage{
combined_dendro_plot(dend, cluster_data, colored_bar, k, colored_bar_label)
}
\description{
Copyright (c) 2019. Kaleido Biosciences. All Rights Reserved
}
