% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmapper.R
\name{heatmapper}
\alias{heatmapper}
\title{heatmapper: Creating an interactive heatmap with colored categorical labels.}
\usage{
heatmapper(phgrofit_output, labels = "Sample.ID")
}
\arguments{
\item{phgrofit_output}{This is the output from phgrofit}

\item{labels}{This is a charachter vector specifiying what colored labels you would like to be displayed beside the heatmap. There can be several labels, but the color palette will get overwhelmed if there are too many labels.}
}
\value{
a plotly heatmap via heatmaply.
}
\description{
heatmapper: Creating an interactive heatmap with colored categorical labels.
}
\examples{
### phgropro processing
phgropro_output = phgropro(biotek_export = filepath.txt,metadata = metadata.csv,Plate_Type = 96)

### phgrofit processing
phgrofit_output = phgrofit(phgropro_output)

### printing heatmap with colored labels for community
community_heatmap = heatmapper(phgrofit_output,"Community")
print(community_heatmap)
}
