% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA.R
\name{PCA}
\alias{PCA}
\title{PCA: Easily making a PCA plot with phgrofit data
Intended to allow the user to easily generate PCA plots from phgrofit data. Note that the values should first be scaled before using this function. This can be acomplished with scale_phgrofit()}
\usage{
PCA(phgrofit_data, group = "Sample.ID", mouse_over = "Compound")
}
\arguments{
\item{phgrofit_data}{This is the output origniating from phgrofit or one of it's modifying functions such as scale_phgrofit or avg_phgrofit.}

\item{group}{This is the name of the column that you would like to color the ordination by. A 95 percent confidence interval for this group will be assigned.}
}
\value{
a ggplot2 object containing the 95 percent confidence interval for the group specified.
}
\description{
PCA: Easily making a PCA plot with phgrofit data
Intended to allow the user to easily generate PCA plots from phgrofit data. Note that the values should first be scaled before using this function. This can be acomplished with scale_phgrofit()
}
\examples{
### phgropro processing
phgropro_output = phgropro(biotek_export = filepath.txt,metadata = metadata.csv,Plate_Type = 96)

### phgrofit processing
phgrofit_output = phgrofit(phgropro_output)

## scaling phgrofit data
phgrofit_data = scale_phgrofit(phgrofit_output)

### printing PCA plot with colored confidence intervals for community
community_PCA = PCA(phgrofit_data,"Community")
print(community_PCA)
}
