% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phgrofit.R
\name{phgrofit}
\alias{phgrofit}
\title{phgrofit: Extract physiological parameters from kinetic pH and OD600 data across diverse samples.}
\usage{
phgrofit(phgropro_output)
}
\arguments{
\item{phgropro_output}{This is the output of phgropro which is a tidy dataframe containing columns for the Sample.ID, any metadata passed to phgropro, and the}
}
\value{
A tidy data frame of 10 values extracted from the spline interpolation.
\itemize{
 \item{"od600_lag_length"}{This is the length of the calculated lag phase.
 Calculated by determining the time where the tangent line at the point of the max growth rate meets the starting od600 }
 \item{"od600_max_gr"}{ This is the maximum growth rate that is observed. Calculated by determining the max derivitive of the spline fit for OD600}
 \item{"max_od600"}{ This is the maximum od600 observed by the spline fit}
 \item{"difference_between_max_and_end_od600"}{ This is the difference between the maximum and end od600. Higher values should correspond to a "death phase". Or one could argue the cells are getting smaller.}
 \item{"max acidification rate"}{ This is the max acidification rate observed in the spline fit. Calculated by determining the min derivitive of the spline fit for pH.}
 \item{"min pH"}{ This is the minimum pH that is observed in the spline fit.}
 \item{"time_of_min_pH "}{ This is the time that the minimum pH occurs in the spline fit.}
 \item{"max_basification_rate"}{ This is the max basification rate observed in the spline fit. Calculated by determining the max derivitive of the spline fit for pH.}
 \item{"max_pH"}{ This is the max pH observed in the spline fit.}
 \item{"difference_between_end_and_min_pH"}{ This is the difference between the end and the minimum pH. A higher value corresponds to a greater pH increase. Higher values may reflect an increased proteolytic state}
 }
}
\description{
phgrofit takes pH and OD600 data that has been formated by phgropro and applies a spline interpolation to extract relevant physiological data.
This is different from the phgrofit modeling because the modeling parameters used are more robust to deviations from an idealized growth and ph profile.
}
\examples{
### phgropro processing
phgropro_output = phgropro(biotek_export = filepath.txt,metadata = metadata.csv,Plate_Type = 96)

### phgrofit processing
phgrofit_output = phgrofit(phgropro_output)
}
