% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phgropro.R
\name{phgropro}
\alias{phgropro}
\title{phgropro: Tidying kinetic od600 and pH data from biotek plate reader.
Copyright (c) 2019. Kaleido Biosciences. All Rights Reserved}
\usage{
phgropro(biotek_export, metadata, Plate_Type = 96)
}
\arguments{
\item{biotek_export}{.txt file that results from a specific format of exporting data from the Biotek Gen5 software.}

\item{metadata}{.csv file that contains information about the contents of the samples. Must contain a column called Sample.ID.}

\item{Plate_Type}{96 or 384 specifying which plate type was ran on the plate reader.}
}
\value{
tidy data frame with a column for Sample ID, any metadata columns present,Time, OD600, and pH.
}
\description{
phgropro- pH growh processing- takes the export from a standardized biotek plate reader and converts it into a tidy format that is convenient for data analysis. This tidy format will oftern serve as the input to phgrofit.
}
